requirements.txt
torch>=2.1
numpy>=1.24
pandas>=2.0
matplotlib>=3.8
scikit-learn>=1.3
pyyaml>=6.0

